Refactor services to further decouple utils and separate out dependecies via uv groups
