+++ assignments/movie-sentiment-aws/src/streamlit_frontend/Dockerfile
@@ -1,35 +1,25 @@
 # Dockerfile for the Streamlit Frontend
 
-# Use a slim, official Python runtime as a parent image
 FROM python:3.12-slim
 
 # Set environment variables for non-interactive installation
 ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
 
-# Set the working directory in the container
 WORKDIR /app
 
-# Install 'uv' for faster dependency management
 RUN pip install uv
 
-# Copy the dependency definition file
-COPY pyproject.toml ./
-
-# Install dependencies using uv
-# This installs all dependencies listed in pyproject.toml
-RUN uv pip install --system --no-cache -r pyproject.toml
-
-# Copy the application source code, config, and assets into the container
 COPY ./src/streamlit_frontend ./src/streamlit_frontend
 COPY ./src/utils ./src/utils
 COPY ./config.yaml ./config.yaml
 COPY ./assets ./assets
+COPY pyproject.toml ./
+
+# Install dependencies using uv that are specific to the frontend
+RUN uv pip install --system --no-cache ".[frontend]"
 
 # Expose the port the app runs on
 EXPOSE 8501
 
-# Define the command to run the application
-# This command starts the Streamlit server.
-# Note: We need to provide the backend URL via an environment variable.
 CMD ["streamlit", "run", "src/streamlit_frontend/app.py"]
